+++ katana/api/response.py
@@ -1,7 +1,7 @@
 """
 Python 3 SDK for the KATANA(tm) Framework (http://katana.kusanagi.io)
 
-Copyright (c) 2016-2017 KUSANAGI S.L. All rights reserved.
+Copyright (c) 2016-2018 KUSANAGI S.L. All rights reserved.
 
 Distributed under the MIT license.
 
@@ -17,7 +17,7 @@ from .http.request import HttpRequest
 from .http.response import HttpResponse
 
 __license__ = "MIT"
-__copyright__ = "Copyright (c) 2016-2017 KUSANAGI S.L. (http://kusanagi.io)"
+__copyright__ = "Copyright (c) 2016-2018 KUSANAGI S.L. (http://kusanagi.io)"
 
 NO_RETURN_VALUE = object()
 
@@ -37,13 +37,8 @@ class Response(Api):
 
     def __init__(self, transport, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        self._logger = RequestLogger(kwargs.get('rid'), 'katana.api')
         self.__attributes = kwargs['attributes']
-
-        # Logging is only enabled when debug is True
-        if self.is_debug():
-            rid = transport.get_request_id()
-            self._logger = RequestLogger(rid, 'katana.api')
-
         self.__gateway_protocol = kwargs.get('gateway_protocol')
         self.__gateway_addresses = kwargs.get('gateway_addresses')