+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.0.0] - 2018-03-01
+### Added
+- Added "log-level" support to runner.
+- Implemented support for the Transport API.
+
+### Changed
+- Removed the "quiet" CLI flag.
+- Logging changed to support the new Syslog based logging from KATANA.
+
 ## [1.3.2] - 2018-01-22
 ### Changed
 - Action schema entity's primary key field name is getted from the